+++ CommonplaceBookApp/ChallengeTemplate.swift
@@ -51,7 +51,7 @@ open class ChallengeTemplate: RawRepresentable {
   /// the template before creating any challenges from it.
   ///
   /// Can only be set from nil to non-nil once; immutable once set.
-  public var templateIdentifier: String?
+  public var templateIdentifier: FlakeID?
 
   /// Subclasses should override and return their particular type.
   /// This is a computed, rather than a stored, property so it does not get serialized.